+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 Release TBD
 - Add orderby field to calendars to control order in selection lists.
 - Fix search results not excluding by calendar status or permissions.
+- Consider calendar perms in Event::hasAccess()
 
 ## v1.4.5
 Release 2018-04-14